% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDataStacked.R
\name{convertDataStacked}
\alias{convertDataStacked}
\title{convertDataStacked}
\usage{
convertDataStacked(data)
}
\arguments{
\item{data}{Either a list of numeric vectors or a data frame where each column
is a vector of observations.}
}
\value{
A stacked data frame.
}
\description{
Creates stacked data frame.
}
\examples{
convertDataStacked(list(x=c(1,2), y=c(2,3,3), z=c(3,4,5,6,7)))
convertDataStacked(data.frame(x=c(1,2), y=c(2,3), z=c(3,4)))
}
