% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kCommonArea.R
\name{kCommonArea}
\alias{kCommonArea}
\title{kCommonArea}
\usage{
kCommonArea(data, bw = bw.nrd0(data[, 1]), npoints = 512)
}
\arguments{
\item{data}{A stacked data frame (see ?convertDataStacked).}

\item{bw}{The bandwidth used to estimate the kernel densities.}

\item{npoints}{The number of points used to estimate the kernel densities.}
}
\value{
Common area between the kernel densities.
}
\description{
Computes the common area between two or more estimed kernel densities.
}
\examples{
data <- convertDataStacked(list(x = rnorm(30), y = rexp(50)))
kCommonArea(data)
}
