% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kTest.R
\name{kSimmetryTest}
\alias{kSimmetryTest}
\title{kSimmetryTest}
\usage{
kSimmetryTest(x, around = "median", ...)
}
\arguments{
\item{x}{A numeric vector of data.}

\item{around}{The desired value to check if the density is simmetryc around. Must be one of 'mean', 'median', or a real number.}

\item{...}{Futher parameters to be passed to kTest.}
}
\value{
A list containing:

- commonArea: Common area between the (data - median) and (median - data) densities.

- p.value: The p-value generated by the permutation test.
}
\description{
Performs a pdf simmetry test for given data based on the estimated kernel densities and the permutation test.
}
\examples{
x = rnorm(100)
ksimmetryTest(x)
}
