% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kTest.R
\name{kTest}
\alias{kTest}
\title{kTest}
\usage{
kTest(data, B = 5000, bw = bw.nrd0(data[, 1]), npoints = 512,
  threads = parallel::detectCores() - 1)
}
\arguments{
\item{data}{A stacked data frame (see ?convertDataStacked).}

\item{B}{Number of permutations.}

\item{bw}{The bandwidth used to estimate the kernel densities.}

\item{npoints}{The number of points used to estimate the kernel densities.}

\item{threads}{Number of cores to be used (see ??DoParallel).}
}
\value{
A list containing:

- commonArea: Common area between the kernel densities.

- p.value: The p-value generated by the permutation test.
}
\description{
Performs a hypothesis test for equality of distributions based on the estimated kernel densities and the permutation test.
}
\examples{
data <- convertDataStacked(list(x = rnorm(30), y = rexp(50)))
kTest(data)
}
