% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kTest.R
\name{simmetryTest}
\alias{simmetryTest}
\title{simmetryTest}
\usage{
simmetryTest(x, ...)
}
\arguments{
\item{x}{A numeric vector of data.}

\item{...}{Futher parameters to be passed to kTest.}
}
\value{
A list containing:

- commonArea: Common area between the (data - median) and (median - data) densities.

- p.value: The p-value generated by the permutation test.
}
\description{
Performs a pdf simmetry test for given data based on the estimated kernel densities and the permutation test.
}
\examples{
x = rnorm(100)
simmetryTest(x)
}
