% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kSymmetryTest.R
\name{kSymmetryTest}
\alias{kSymmetryTest}
\title{kSymmetryTest}
\usage{
kSymmetryTest(x, around = "median", ...)
}
\arguments{
\item{x}{A numeric vector of data.}

\item{around}{The desired value to check if the density is symmetric around. Must be one of 'mean', 'median', or a real number.}

\item{...}{Futher parameters to be passed to \link[ktest]{kTest}.}
}
\value{
A list containing:

- densities: The estimated densities for each group.

- labels: The group labels.

- ca: Common area between the (data - point) and (point - data) densities.

- pvalue: The p-value generated by the permutation test.
}
\description{
Performs a pdf symmetry test for given data based on the estimated kernel densities and the permutation test.
}
\examples{
x = rnorm(100)
test = kSymmetryTest(x)
print(test)
plot(test)
}
