% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kTest.R
\name{kTest}
\alias{kTest}
\title{kTest}
\usage{
kTest(data, classes = NULL, perm = TRUE, B = 5000,
  pairwise = FALSE, bw = bw.nrd0, npoints = 512,
  threads = detectCores() - 1)
}
\arguments{
\item{data}{Either a list of numeric vectors, a numeric vector (with classes parameter defined), or a stacked data frame (first column with numeric values and second column with classes).}

\item{classes}{Classes relative to data parameter, should be used only when data is a numeric vector.}

\item{perm}{Boolean indicating weather to obtain the p-value trough the permutation test or just return return the common area between densities.}

\item{B}{Number of permutations.}

\item{pairwise}{Boolean indicating weather to obtain pairwise p-values (valid only if > 2 groups). If the number of groups is big, this will make the execution much slower.}

\item{bw}{The bandwidth used to estimate the kernel densities.}

\item{npoints}{The number of points used to estimate the kernel densities.}

\item{threads}{Number of cores to be used for parallel computing.}
}
\value{
A list containing:

- densities: The estimated densities for each group.

- labels: The group labels.

- ca: Common area between the kernel densities.

- pvalue: The p-value generated by the permutation test (if perm = TRUE).

- pairwiseca: Groups pairwise common area triangular matrix (just if >= 3 groups).

- pairwisepvalue: Groups pairwise p-value triangular matrix generated by the permutation test (if perm = TRUE, pairwise = TRUE and >= 3 groups).
}
\description{
Performs a hypothesis test for equality of distributions based on the estimated kernel densities and the permutation test.
}
\examples{
data = list(x = rnorm(30), y = rexp(50), z = rpois(70, 1))
test = kTest(data)
print(test)
plot(test)
pairs(test)
}
